misc: refacto like system
